feat: Move traced and measured method filter parsing into config
