+++ dd-java-agent/instrumentation/graal/native-image/src/main/java/datadog/trace/instrumentation/graal/nativeimage/NativeImageGeneratorRunnerInstrumentation.java
@@ -62,6 +62,7 @@ public final class NativeImageGeneratorRunnerInstrumentation
               + "datadog.trace.api.InstrumenterConfig:build_time,"
               + "datadog.trace.api.Functions:build_time,"
               + "datadog.trace.api.GlobalTracer:build_time,"
+              + "datadog.trace.api.MethodFilterConfigParser:build_time,"
               + "datadog.trace.api.WithGlobalTracer:build_time,"
               + "datadog.trace.api.PropagationStyle:build_time,"
               + "datadog.trace.bootstrap.config.provider.ConfigConverter:build_time,"
@@ -78,7 +79,6 @@ public final class NativeImageGeneratorRunnerInstrumentation
               + "datadog.trace.bootstrap.instrumentation.java.concurrent.ExcludeFilter:build_time,"
               + "datadog.trace.bootstrap.instrumentation.java.concurrent.QueueTimeHelper:build_time,"
               + "datadog.trace.bootstrap.instrumentation.java.concurrent.TPEHelper:build_time,"
-              + "datadog.trace.bootstrap.instrumentation.traceannotation.TraceAnnotationConfigParser:build_time,"
               + "datadog.trace.logging.LoggingSettingsDescription:build_time,"
               + "datadog.trace.logging.simplelogger.SLCompatFactory:build_time,"
               + "datadog.trace.util.CollectionUtils:build_time,"
